minor: rename sim_macro_cache.py to simulation_macro_cache.py
fixes: issue #265
